+++ src/hooks/useField.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useContext, useMemo, useRef } from 'react';
+import React, { useState, useEffect, useContext, useRef, useMemo } from 'react';
 import { FormRegisterContext, MultistepStepContext } from '../Context';
 import useFormApi from './useFormApi';
 import useStateWithGetter from './useStateWithGetter';
@@ -185,6 +185,7 @@ function useField(fieldProps = {}, userRef) {
     required,
     keepStateIfRelevant,
     initialize,
+    formatterDependencies = [],
     ...userProps
   } = fieldProps;
 
@@ -210,6 +211,18 @@ function useField(fieldProps = {}, userRef) {
   // Create ref to fieldObject
   const fieldObjectRef = useRef();
 
+  // Create ref for pristine and dirty
+  const valueTouched = useRef(false);
+
+  // Getters for value ref
+  const getPristine = () => {
+    return !valueTouched.current;
+  };
+
+  const getDirty = () => {
+    return valueTouched.current;
+  };
+
   // If the form Controller was passed in then use that instead
   if (formController) {
     updater = formController.updater;
@@ -227,7 +240,7 @@ function useField(fieldProps = {}, userRef) {
   const [formInitialValue] = useState(() => updater.getInitialValue(field));
 
   // We might have keep state so check for it!
-  const savedState = formApi.getSavedValue(field);
+  const savedState = updater.getSavedValue(field);
 
   // Create Initial Values
   let initVal;
@@ -355,6 +368,9 @@ function useField(fieldProps = {}, userRef) {
   const setValue = (v, e, options = {}) => {
     let val = v;
 
+    // This value has now been modified
+    valueTouched.current = true;
+
     logger(`Setting ${field} to ${val}`);
 
     // Set value may have been called externally
@@ -446,6 +462,13 @@ function useField(fieldProps = {}, userRef) {
     updater.setValue(fieldId, val, !options.preventUpdate);
   };
 
+  useEffect(
+    () => {
+      fieldApiRef.current.setValue(fieldApiRef.current.getValue() || '');
+    },
+    [...formatterDependencies]
+  );
+
   // ---- Define set touched ----
   const setTouched = (val, reset, { preventUpdate } = {}) => {
     logger(`Field ${field} has been touched`);
@@ -523,6 +546,9 @@ function useField(fieldProps = {}, userRef) {
       preventUpdate
     });
     setTouched(undefined, true, { preventUpdate });
+
+    // We are now at our initial state
+    valueTouched.current = false;
   };
 
   // ---- Define validate ----
@@ -558,9 +584,13 @@ function useField(fieldProps = {}, userRef) {
     getTouched: getTouch,
     getError: getErr,
     getIsRelevant: getIsRelevant,
+    getDirty,
+    getPristine,
     getFieldState: () => ({
       value: getVal(),
-      touched: getTouch()
+      touched: getTouch(),
+      dirty: getDirty(),
+      pristine: getPristine()
     }),
     relevant,
     multistepRelevant,
@@ -574,7 +604,9 @@ function useField(fieldProps = {}, userRef) {
     error,
     touched,
     maskedValue,
-    isRelevant
+    isRelevant,
+    dirty: valueTouched.current,
+    pristine: !valueTouched.current
   };
 
   // Create shadow state if this is a shadow field
@@ -658,6 +690,11 @@ function useField(fieldProps = {}, userRef) {
 
   const render = children =>
     useMemo(() => (isRelevant ? children : null), [...shouldUpdate]);
+
+  // const render = children => {
+  //   // console.log('RENDER', field);
+  //   return isRelevant ? children : null;
+  // };
 
   // Build some setub fields so users can easily intagrate without any hookup code